Replace PMCR with get_ticks() for network timing, use RTC for DHCP lease tracking
Route all network timeout loops through target_ops_t get_ticks() (now uint64_t)
instead of direct PMCR access. Use compare-and-increment pattern to avoid
__udivdi3 (~1KB libgcc). Track DHCP lease expiry via RTC so countdown survives
program execution.

- get_ticks() returns uint64_t: GC reads full 64-bit TBR, DC zero-extends TMU2
- Add get_rtc() to target_ops_t for both DC and GC
- Replace PMCR_Read + 64-bit division in 5 adapter drivers with tick comparison
- Replace get_some_time() in both dhcp.c with get_ticks() for XID entropy
- Rewrite lease tracking: RTC-based expiry + tick-based ~1Hz display countdown
- Resync lease from RTC in disp_info() so it updates after program return
- Remove PMCR_Init from network_transport.c, perfctr.h from all network files
- Move cdfs_init() from common main.c to DC target init

+++ client/common/screensaver.c
@@ -36,8 +36,8 @@ static const uint32_t screensaver_icon[32] = {
 
 /* State */
 static void (*restore_callback)(void);
-static unsigned int timer_start;
-static unsigned int last_frame;
+static uint64_t timer_start;
+static uint64_t last_frame;
 static int active;
 static unsigned int icon_color_saved;
 static int box_x, box_y;
@@ -85,11 +85,11 @@ int screensaver_wake(void)
 void screensaver_poll(void)
 {
     const target_ops_t *t = common_get_target();
-    unsigned int now = t->get_ticks();
+    uint64_t now = t->get_ticks();
 
     if (!active) {
         /* Check if idle timeout has elapsed (unsigned subtraction handles wrap) */
-        unsigned int elapsed = now - timer_start;
+        uint64_t elapsed = now - timer_start;
         unsigned int timeout_ticks = IDLE_TIMEOUT_SECS * t->ticks_per_second;
         if (elapsed < timeout_ticks)
             return;
